requirements.txt
torch
torchvision
torchaudio
numpy
pillow
opencv-python
grad-cam==1.4.8
gradio>=3.37.0
ultralytics
torchcam
